requirements.txt
streamlit
tensorflow
pillow
numpy
tf_keras
tensorflow_hub